feat(#S11P12D203-162): 점령지 현황 API 연결

+++ android/project/app/src/main/java/com/chill/mallang/data/model/entity/Teams.kt
@@ -11,5 +11,5 @@ data class TeamList(
 data class Team(
     val teamName: String,
     val area: Int,
-    val teamId: Long
+    val teamId: Int,
 )
